History and profile data tier and application tier

+++ Elearn-ApplicationTier/Elearn.Application/LogicContracts/ILectureLogic.cs
@@ -8,6 +8,6 @@ public interface ILectureLogic
     Task<Lecture> CreateAsync(Lecture lecture);
     Task<List<Lecture>> GetAllLecturesAsync();
     Task<Lecture?> GetLectureAsync(string url);
-    Task<List<Lecture>> GetLectureByUserIdAsync(int userId);
-    Task<List<Lecture>> GetUpvotedLectureByUserIdAsync(int userId);
+    Task<List<Lecture>> GetLectureByUserIdAsync(long userId);
+    Task<List<Lecture>> GetUpvotedLectureByUserIdAsync(long userId);
 }
