+++ Elearn-ApplicationTier/Elearn.Application/ServiceContracts/IUserService.cs
@@ -6,7 +6,7 @@ public interface IUserService
 {
     Task<User?> GetUserByNameAsync(string name);
     Task<User> CreateNewUserAsync(User user);
-     Task<User?> GetUserByIdAsync(int id);
+     Task<User?> GetUserByIdAsync(long id);
     Task<User> UpdateUserAsync(User updated);
     Task<User?> GetUserByUsernameAsync(string username);
 }
